+++ src/config/global-config.c
@@ -21,7 +21,7 @@ int Read_GlobalSK(XML_NODE node, void *configp, __attribute__((unused)) void *ma
     unsigned int ign_size = 1;
     const char *xml_ignore = "ignore";
     const char *xml_auto_ignore = "auto_ignore";
-    const char *xml_ignore_frecuency = "frecuency";
+    const char *xml_ignore_frequency = "frequency";
     const char *xml_ignore_time = "timeframe";
     const char *xml_alert_new_files = "alert_new_files";
 
@@ -51,7 +51,7 @@ int Read_GlobalSK(XML_NODE node, void *configp, __attribute__((unused)) void *ma
             merror(XML_VALUENULL, node[i]->element);
             return (OS_INVALID);
         } else if (strcmp(node[i]->element, xml_auto_ignore) == 0) {
-            Config->syscheck_ignore_frecuency = 10;
+            Config->syscheck_ignore_frequency = 10;
             Config->syscheck_ignore_time = 3600;
             if (strcmp(node[i]->content, "yes") == 0) {
                 Config->syscheck_auto_ignore = 1;
@@ -62,13 +62,13 @@ int Read_GlobalSK(XML_NODE node, void *configp, __attribute__((unused)) void *ma
                 return (OS_INVALID);
             }
             for (j = 0; node[i]->attributes && node[i]->attributes[j]; ++j) {
-                if (strcmp(node[i]->attributes[j], xml_ignore_frecuency) == 0) {
+                if (strcmp(node[i]->attributes[j], xml_ignore_frequency) == 0) {
                     if (!OS_StrIsNum(node[i]->values[0])) {
                         merror(XML_VALUEERR, node[i]->attributes[j], node[i]->values[j]);
                         return (OS_INVALID);
                     }
-                    Config->syscheck_ignore_frecuency = atoi(node[i]->values[0]);
-                    if (Config->syscheck_ignore_frecuency < 1 || Config->syscheck_ignore_frecuency > 99) {
+                    Config->syscheck_ignore_frequency = atoi(node[i]->values[0]);
+                    if (Config->syscheck_ignore_frequency < 1 || Config->syscheck_ignore_frequency > 99) {
                         merror(XML_VALUEERR, node[i]->attributes[j], node[i]->values[j]);
                         return (OS_INVALID);
                     }